// petforest-fe-cms/src/utils/enums/post/post.enum.ts
export enum POST_ENUM_TABS {
  POST = 'POST',
  REVIEW = 'REVIEW',
  POPUP = 'POPUP',
}

export enum POST_MAIN_CATEGORIES {
  ANNOUNCEMENT = 'ANNOUNCEMENT',
  FREQUENTLY_ASKED = 'FREQUENTLY_ASKED',
  PRESS_REPORT = 'PRESS_REPORT',
  JOURNAL = 'JOURNAL',
}

export enum POST_CATEGORIES {
  ANNOUNCEMENT = 'ANNOUNCEMENT',
  FREQUENTLY_ASKED = 'FREQUENTLY_ASKED',
  FREQUENTLY_ASKED_BASIC_TRAINING = 'FREQUENTLY_ASKED_BASIC_TRAINING',
  FREQUENTLY_ASKED_ETC = 'FREQUENTLY_ASKED_ETC',
  FREQUENTLY_ASKED_LUCETTE = 'FREQUENTLY_ASKED_LUCETTE',
  FREQUENTLY_ASKED_FUNERAL_SERVICE = 'FREQUENTLY_ASKED_FUNERAL_SERVICE',
  FREQUENTLY_ASKED_FUNERAL_MATERIAL = 'FREQUENTLY_ASKED_FUNERAL_MATERIAL',
  JOURNAL = 'JOURNAL',
  JOURNAL_HEALTH = 'JOURNAL_HEALTH',
  JOURNAL_EDUCATION = 'JOURNAL_EDUCATION',
  JOURNAL_ETC = 'JOURNAL_ETC',
  JOURNAL_CULTURAL = 'JOURNAL_CULTURAL',
  // JOURNAL_LAW = 'JOURNAL_LAW',
  JOURNAL_FUNERAL = 'JOURNAL_FUNERAL',
  JOURNAL_PETRO = 'JOURNAL_PETRO',
  PRESS_REPORT = 'PRESS_REPORT',
}

export enum POST_VISIBILITY {
  RELEASE = 'RELEASE',
  PRIVATE = 'PRIVATE',
}

export enum POST_EXPOSURE {
  EXPOSED = 'EXPOSED',
  UNEXPOSED = 'UNEXPOSED',
}